Modify seeds file for development
